fix cross repo build errs
