+++ packages/mongodb-chatbot-server/src/processors/index.ts
@@ -1,5 +1,9 @@
+export * from "./FindContentFunc";
+export * from "./GenerateUserPromptFunc";
 export * from "./makeBoostOnAtlasSearchFilter";
+export * from "./makeDefaultFindContent";
+export * from "./makeDefaultReferenceLinks";
+export * from "./makeRagGenerateUserPrompt";
+export * from "./MakeReferenceLinksFunc";
 export * from "./QueryPreprocessorFunc";
 export * from "./SearchBooster";
-export * from "./GenerateUserPromptFunc";
-export * from "./makeRagGenerateUserPrompt";
