+++ packages/mongodb-chatbot-server/src/processors/makeDefaultReferenceLinks.test.ts
@@ -1,0 +1,119 @@
+import { ObjectId, EmbeddedContent } from "mongodb-rag-core";
+import { makeDefaultReferenceLinks } from "./makeDefaultReferenceLinks";
+
+describe("makeDefaultReferenceLinks()", () => {
+  // Chunk 1 and 2 are the same page. Chunk 3 is a different page.
+  const chunk1 = {
+    _id: new ObjectId(),
+    url: "https://mongodb.com/docs/realm/sdk/node/",
+    text: "blah blah blah",
+    tokenCount: 100,
+    embedding: [0.1, 0.2, 0.3],
+    updated: new Date(),
+    sourceName: "realm",
+  };
+  const chunk2 = {
+    _id: new ObjectId(),
+    url: "https://mongodb.com/docs/realm/sdk/node/",
+    text: "blah blah blah",
+    tokenCount: 100,
+    embedding: [0.1, 0.2, 0.3],
+    updated: new Date(),
+    sourceName: "realm",
+  };
+  const chunk3 = {
+    _id: new ObjectId(),
+    url: "https://mongodb.com/docs/realm/sdk/node/xyz",
+    text: "blah blah blah",
+    tokenCount: 100,
+    embedding: [0.1, 0.2, 0.3],
+    updated: new Date(),
+    sourceName: "realm",
+  };
+  const chunkWithTitle = {
+    _id: new ObjectId(),
+    url: "https://mongodb.com/docs/realm/sdk/node/xyz",
+    text: "blah blah blah",
+    metadata: {
+      pageTitle: "title",
+    },
+    tokenCount: 100,
+    embedding: [0.1, 0.2, 0.3],
+    updated: new Date(),
+    sourceName: "realm",
+  };
+  test("No sources should return empty string", () => {
+    const noChunks: EmbeddedContent[] = [];
+    const noReferences = makeDefaultReferenceLinks(noChunks);
+    expect(noReferences).toEqual([]);
+  });
+  test("One source should return one link", () => {
+    const oneChunk: EmbeddedContent[] = [chunk1];
+    const oneReference = makeDefaultReferenceLinks(oneChunk);
+    const expectedOneReference = [
+      {
+        title: "https://mongodb.com/docs/realm/sdk/node/",
+        url: "https://mongodb.com/docs/realm/sdk/node/",
+      },
+    ];
+    expect(oneReference).toEqual(expectedOneReference);
+  });
+  test("Chunk with title should return title in reference", () => {
+    const oneChunk: EmbeddedContent[] = [chunkWithTitle];
+    const oneReference = makeDefaultReferenceLinks(oneChunk);
+    const expectedOneReference = [
+      {
+        title: "title",
+        url: "https://mongodb.com/docs/realm/sdk/node/xyz",
+      },
+    ];
+    expect(oneReference).toEqual(expectedOneReference);
+  });
+  test("Multiple sources from same page should return one link", () => {
+    const twoChunksSamePage: EmbeddedContent[] = [chunk1, chunk2];
+    const oneReferenceSamePage = makeDefaultReferenceLinks(twoChunksSamePage);
+    const expectedOneReferenceSamePage = [
+      {
+        title: "https://mongodb.com/docs/realm/sdk/node/",
+        url: "https://mongodb.com/docs/realm/sdk/node/",
+      },
+    ];
+    expect(oneReferenceSamePage).toEqual(expectedOneReferenceSamePage);
+  });
+  test("Multiple sources from different pages should return 1 link per page", () => {
+    const twoChunksDifferentPage: EmbeddedContent[] = [chunk1, chunk3];
+    const multipleReferencesDifferentPage = makeDefaultReferenceLinks(
+      twoChunksDifferentPage
+    );
+    const expectedMultipleReferencesDifferentPage = [
+      {
+        title: "https://mongodb.com/docs/realm/sdk/node/",
+        url: "https://mongodb.com/docs/realm/sdk/node/",
+      },
+      {
+        title: "https://mongodb.com/docs/realm/sdk/node/xyz",
+        url: "https://mongodb.com/docs/realm/sdk/node/xyz",
+      },
+    ];
+    expect(multipleReferencesDifferentPage).toEqual(
+      expectedMultipleReferencesDifferentPage
+    );
+    // All three sources. Two from the same page. One from a different page.
+    const threeChunks: EmbeddedContent[] = [chunk1, chunk2, chunk3];
+    const multipleSourcesWithSomePageOverlap =
+      makeDefaultReferenceLinks(threeChunks);
+    const expectedMultipleSourcesWithSomePageOverlap = [
+      {
+        title: "https://mongodb.com/docs/realm/sdk/node/",
+        url: "https://mongodb.com/docs/realm/sdk/node/",
+      },
+      {
+        title: "https://mongodb.com/docs/realm/sdk/node/xyz",
+        url: "https://mongodb.com/docs/realm/sdk/node/xyz",
+      },
+    ];
+    expect(multipleSourcesWithSomePageOverlap).toEqual(
+      expectedMultipleSourcesWithSomePageOverlap
+    );
+  });
+});
